Add option to use floating time windows
When issue #269 was fixed, the requested "window" (start to end
time) was changed to be aligned to 5 minute slots. This turned
out to work badly when using the metrics as alarms, as a request
at e.g. 12:38:31 with length 300s would use a window of 12:35:00
to 12:40:00. This would mean that the actual length (300s) would
not be respected, and that any requests does use its own "dynamic
delay" to round the time to 5 min intervals.

This patch adds an option to go back to the old behaviours of
windows that are not fixed to the 5 min slots.