Add support for specifying an External ID value in IAM role ARNs.
This is a breaking change as it requires the YAML config file to be
updated if you're currently using the `roleArns` field.

**Context**
IAM role delegation allows optionally setting an External ID string
value shared between parties. This is useful and recommended in order to
prevent the "confused deputy" problem when the account that can assume
the IAM role is outside your organisation.
When an External ID is setup for a role, users in the trusted account
must provide the exact identifier value to be able to assume the role.

More context here: https://docs.aws.amazon.com/IAM/latest/UserGuide/id_roles_create_for-user_externalid.html

**Changes**
In this PR we change the config file format to allow specifying IAM
roles with both an ARN string and an External ID string. The credentials
handling code is updated to check for the presence of the External ID
value and pass it along in the assumeRole request, if present.