Allow setting default statistics in job config
We can already set default values for most values, but we can't for
statistics, and force the user to explicitly set them. I cannot
immediately see why we can't also set statistics with a default value.

This copies the logic for other values, but instead of setting a
default, it fails if neither a default or an explicit value is set.