Fix searchTags and bad dimension name
This commit resolves https://github.com/nerdswords/yet-another-cloudwatch-exporter/issues/365.

In cases in which the name of the dimensions for a given metric didn't
match a dimension in our searchTags list, we would pass a default global query
for all metrics under a given namespace into our list, returning metrics
we meant to filter. A test case was added to cover.

In addition, I decorated a debug log entry that was useful in finding
this bug.