New max-dimensions associator (feature flagged)
Introduce a new algorithm for auto-discovery jobs to match ListMetrics
names and dimensions to tagged resources. The algorithm is an extension
of the current one, it takes into account as much dimensions as possible
and makes sure that resources are mapped only once.

It is currently opt-in, can be enabled via a `max-dimensions-associator`
feature flag.