Fix AmazonMQ Broker name dimension match.
For active/standby ActiveMQ brokers (not for RabbitMQ), the value of the
"Broker" dimension contains a number suffix that is not part of the
resource ARN. This causes a mismatch with the resource ARN, which
contains the broker name without any suffix instead (there's only one
ARN per active/standby group).

Add as special case to handle this in the new max-dim-associator.