Limit number of goroutines for GetMetricData calls
The CloudWatch client already limits the number of concurrent requests
to GetMetricData API via the `-cloudwatch-concurrency` flag, however
in discovery jobs there's no limit to how many goroutines can be spawn
at the same time and get blocked waiting for the client to be available.

With this change, we use an errgroup to block adding goroutines to
handle GetMetricData partitions until the client is actually available.