scrape: Return resources even if there are no metrics
If no samples have been recorded for a metric in a region, for example
if the resource is new and hasn't been used yet, CloudWatch will not
return the metrics via `ListMetrics`. In that case, we discard any
resources we discovered and don't emit `_info` metrics for them.

But it might still be useful to know your entire estate even if some of
the resources don't have metrics yet. So let's still emit the resources
even when there are no metrics for them.