main: Initialise logger when exiting if needed
If we exit very early, inside `app.Run` itself, we can get a panic.

This happens when we're given an unknown commandline flag, for example.
In that case  `urfave/cli` exits  before parsing the remaining flags
and before we could have a chance to set the logger up from them.

When exiting, check if the logger is `nil`, and create one with default
parameters if it is.