Change config Validate() signature to include model conversion
In a previous PR I've changed the Load() signature to include the
conversion of the yaml-based config into internal representation. This
PR changes the Validate() function as well so that library users can
validate and convert an in-memory config in one pass.

Followup of https://github.com/nerdswords/yet-another-cloudwatch-exporter/pull/1161