Refactor dimensions regexp usage for discovery jobs
In discovery jobs, the associator extracts dimensions names from
services regexp definitions at each scrape. With this changes, names
are pre-computed as part of building the job model. This also fixes
a race condition where the slice of dimensions names was modified in
place (whereas regexp.SubexpNames() suggests it shouldn't).