Restructure fields on CloudwatchData (#1289)
This PR restructures some of the fields on the CloudwatchData struct to make the field usage much more obvious. This struct is used through out the scraping process to carry around context and the result of finding metric data and contains a lot of overlapping/overloaded fields. The net result should be a struct that is much easier to understand making some of the complex mapping done in migrate.BuildMetrics a little more obvious, and less pointers where they are not necessary.

Related to: #1290